Move builder to its own file for better clarity
